Cli 4 Sin servicios

+++ CLIENTS/Logic/Manager/ClientManager.cs
@@ -2,6 +2,9 @@
 using System.Collections.Generic;
 using UPB.ProyectoFinal.Logic.Model;
 using UPB.ProyectoFinal.Data;
+using UPB.ProyectoFinal.Logic.Exceptions;
+using Serilog;
+
 
 namespace UPB.ProyectoFinal.Logic.Manager
 {
@@ -14,15 +17,50 @@ namespace UPB.ProyectoFinal.Logic.Manager
             _dbContext = dbContext;
         }
 
+        
         public List<Client> CreateClient(Client client)
         {
-            client.Codigo = GenerateCode(client);
-             
+           if (client.Nombre =="") {
+                ClientAttrException e = new ClientAttrException("No puede dejar el nombre vacio");
+                Log.Error("Se produjo un error"+e.StackTrace+e.Message);
+                throw e;
+            }
+            if (!(10000000 < client.CI && client.CI < 100000000))
+            {
+                ClientAttrException e = new ClientAttrException("Carnet de identidad invalido");
+                Log.Error("Se produjo un error" + e.StackTrace + e.Message);
+                throw e;
+            }
+            if (!(0<client.Ranking && client.Ranking<6))
+            {
+                ClientAttrException e = new ClientAttrException("Ranking es invalido");
+                Log.Error("Se produjo un error" + e.StackTrace + e.Message);
+                throw e;
+            }
+            client.Codigo = GenerateCode(client);  
             return DTOMappers.MapClients(_dbContext.CreateClient(DTOMappers.MapClientLD(client)));
         }
 
         public List<Client> DeleteClient(Client client)
         {
+            if (client.Nombre == "")
+            {
+                ClientAttrException e = new ClientAttrException("No puede dejar el nombre vacio");
+                Log.Error("Se produjo un error" + e.StackTrace + e.Message);
+                throw e;
+            }
+            if (!(10000000 < client.CI && client.CI < 100000000))
+            {
+                ClientAttrException e = new ClientAttrException("Carnet de identidad invalido");
+                Log.Error("Se produjo un error" + e.StackTrace + e.Message);
+                throw e;
+            }
+            if (!(10000000 < client.CI && client.CI < 100000000))
+            {
+                ClientAttrException e = new ClientAttrException("Ranking es invalido");
+                Log.Error("Se produjo un error" + e.StackTrace + e.Message);
+                throw e;
+            }
             return DTOMappers.MapClients(_dbContext.DeleteClient(DTOMappers.MapClientLD(client)));
         }
 
@@ -33,6 +71,25 @@ namespace UPB.ProyectoFinal.Logic.Manager
 
         public List<Client> UpdateClient(Client client)
         {
+            if (client.Nombre == "")
+            {
+                ClientAttrException e = new ClientAttrException("No puede dejar el nombre vacio");
+                Log.Error("Se produjo un error" + e.StackTrace + e.Message);
+                throw e;
+            }
+            if (!(10000000 < client.CI && client.CI < 100000000))
+            {
+                ClientAttrException e = new ClientAttrException("Carnet de identidad invalido");
+                Log.Error("Se produjo un error" + e.StackTrace + e.Message);
+                throw e;
+            }
+            if (!(10000000 < client.CI && client.CI < 100000000))
+            {
+                ClientAttrException e = new ClientAttrException("Ranking es invalido");
+                Log.Error("Se produjo un error" + e.StackTrace + e.Message);
+                throw e;
+            }
+            client.Codigo = GenerateCode(client);
             return DTOMappers.MapClients(_dbContext.UpdateClient(DTOMappers.MapClientLD(client)));
         }
